requirements.txt
torch==2.0.1
numpy==1.24.4
torchvision==0.15.2
gym==0.26.1
gym[atari]==0.26.1
gym[accept-rom-license]==0.26.1
cython
PyOpenGL-accelerate
hydra-core
protobuf==3.20.1
wandb==0.14.2
dm_control
opencv-python
kornia
imageio==2.31.5
imageio[ffmpeg]
imageio[pyav]
ray==1.0.0
torchrl==0.1.0
git+https://gitclone.com/github.com/denisyarats/dmc2gym.git
tensordict==0.2.0
pillow>=5.3.0
colorednoise==2.2.0
matplotlib==3.3.4
scikit-image==0.18.3
gymnasium==0.26.3